# server/swagger.yaml
openapi: 3.1.0
info:
  title: Library API
  description: API quản lý thư viện sách trực tuyến, hỗ trợ Chủ sách và Người đọc
    theo chuẩn OpenAPI
  version: 1.0.0
servers:
- url: http://localhost:3000
  description: Development Server
paths:
...
  /api/v1/books:
    get:
      tags:
      - Books
      summary: Get Books
      description: Tìm kiếm và lấy danh sách sách
      operationId: get_books_api_v1_books_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Từ khóa tìm kiếm chung (tùy chọn)
          title: Q
        description: Từ khóa tìm kiếm chung (tùy chọn)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_BookResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Books
      summary: Create Book
      description: Đăng sách mới
      operationId: create_book_api_v1_books_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/books/{id}:
    get:
      tags:
      - Books
      summary: Get Book
      description: Xem chi tiết cuốn sách
      operationId: get_book_api_v1_books__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Books
      summary: Update Book
      description: Sửa thông tin sách (yêu cầu quyền Chủ sách)
      operationId: update_book_api_v1_books__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Books
      summary: Delete Book
      description: Xóa sách (yêu cầu quyền Chủ sách)
      operationId: delete_book_api_v1_books__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/books/admin/force-delete/{id}:
    delete:
      tags:
      - Books
      summary: Admin Force Delete Book
      description: Xóa sách bất kỳ không cần hỏi (Chỉ dành cho Admin)
      operationId: admin_force_delete_book_api_v1_books_admin_force_delete__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/borrows:
    post:
      tags:
      - Borrow Records
      summary: Create Borrow
      description: Tạo đơn mượn (truyền book_id hoặc collection_id)
      operationId: create_borrow_api_v1_borrows_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRecordCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Borrow Records
      summary: Get Borrows
      description: Lấy danh sách lịch mượn tài khoản đang tham gia (đơn đi mượn hoặc
        đơn người khác mượn sách của mình)
      operationId: get_borrows_api_v1_borrows_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_BorrowRecordResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/borrows/{id}:
    get:
      tags:
      - Borrow Records
      summary: Get Borrow
      description: Xem chi tiết một lịch mượn cụ thể
      operationId: get_borrow_api_v1_borrows__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/borrows/{id}/status:
    patch:
      tags:
      - Borrow Records
      summary: Update Borrow Status
      description: Cập nhật trạng thái duyệt/từ chối hoặc trả sách (dành cho chủ sách/người
        mượn)
      operationId: update_borrow_status_api_v1_borrows__id__status_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRecordUpdateStatus'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/books:
    get:
      tags:
      - Books
      summary: Get Books
      description: Tìm kiếm và lấy danh sách sách
      operationId: get_books_api_v2_books_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Từ khóa tìm kiếm chung (tùy chọn)
          title: Q
        description: Từ khóa tìm kiếm chung (tùy chọn)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_BookResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Books
      summary: Create Book
      description: Đăng sách mới
      operationId: create_book_api_v2_books_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/books/{id}:
    get:
      tags:
      - Books
      summary: Get Book
      description: Xem chi tiết cuốn sách
      operationId: get_book_api_v2_books__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Books
      summary: Update Book
      description: Sửa thông tin sách (yêu cầu quyền Chủ sách)
      operationId: update_book_api_v2_books__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Books
      summary: Delete Book
      description: Xóa sách (yêu cầu quyền Chủ sách)
      operationId: delete_book_api_v2_books__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/books/admin/force-delete/{id}:
    delete:
      tags:
      - Books
      summary: Admin Force Delete Book
      description: Xóa sách bất kỳ không cần hỏi (Chỉ dành cho Admin)
      operationId: admin_force_delete_book_api_v2_books_admin_force_delete__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/borrows:
    post:
      tags:
      - Borrow Records
      summary: Create Borrow
      description: Tạo đơn mượn (truyền book_id hoặc collection_id)
      operationId: create_borrow_api_v2_borrows_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRecordCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Borrow Records
      summary: Get Borrows
      description: Lấy danh sách lịch mượn tài khoản đang tham gia (đơn đi mượn hoặc
        đơn người khác mượn sách của mình)
      operationId: get_borrows_api_v2_borrows_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_BorrowRecordResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/borrows/{id}:
    get:
      tags:
      - Borrow Records
      summary: Get Borrow
      description: Xem chi tiết một lịch mượn cụ thể
      operationId: get_borrow_api_v2_borrows__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/borrows/{id}/status:
    patch:
      tags:
      - Borrow Records
      summary: Update Borrow Status
      description: Cập nhật trạng thái duyệt/từ chối hoặc trả sách (dành cho chủ sách/người
        mượn)
      operationId: update_borrow_status_api_v2_borrows__id__status_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRecordUpdateStatus'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRecordResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BookCreate:
      properties:
        title:
          type: string
          title: Title
        author:
          type: string
          title: Author
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quantity
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
      type: object
      required:
      - title
      - author
      title: BookCreate
    BookResponse:
      properties:
        title:
          type: string
          title: Title
        author:
          type: string
          title: Author
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quantity
        id:
          type: integer
          title: Id
        owner_id:
          type: integer
          title: Owner Id
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - title
      - author
      - id
      - owner_id
      - collection_id
      - created_at
      title: BookResponse
    BookUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quantity
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
      type: object
      title: BookUpdate
    BorrowRecordCreate:
      properties:
        book_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Book Id
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
        additional_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Additional Info
      type: object
      title: BorrowRecordCreate
    BorrowRecordResponse:
      properties:
        id:
          type: integer
          title: Id
        book_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Book Id
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
        borrower_id:
          type: integer
          title: Borrower Id
        owner_id:
          type: integer
          title: Owner Id
        status:
          $ref: '#/components/schemas/BorrowStatus'
        additional_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Additional Info
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - book_id
      - collection_id
      - borrower_id
      - owner_id
      - status
      - additional_info
      - created_at
      title: BorrowRecordResponse
    BorrowRecordUpdateStatus:
      properties:
        status:
          $ref: '#/components/schemas/BorrowStatus'
      type: object
      required:
      - status
      title: BorrowRecordUpdateStatus
    BorrowStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - returned
      - expired
      title: BorrowStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PageResponse_BookResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BookResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        total_pages:
          type: integer
          title: Total Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - total_pages
      title: PageResponse[BookResponse]
    PageResponse_BorrowRecordResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BorrowRecordResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        total_pages:
          type: integer
          title: Total Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - total_pages
      title: PageResponse[BorrowRecordResponse]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login